# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaPhotoAPI Specification
  description: |
    This OpenAPI document describes the WasaPhoto project REST API.

    WasaPhoto is a social media platform that focuses on photo sharing.

  version: 1.0.0
servers:
  - url: "http://localhost:3000"
    description: localhost server for development
security:
  - bearerAuthentication: []
...
paths:
...
  /posts/{postId}:
    delete:
      tags: ["post"]
      operationId: deletePhoto
      summary: Remove a post.
      description: |-
        Remove a post.
      parameters: 
        - $ref: "#/components/parameters/PostId"
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /posts/{postId}/likes:
    post:
      tags: ["post"]
      operationId: likePhoto
      summary: Like a post.
      description: |-
        Like a post.
      parameters: 
        - $ref: "#/components/parameters/PostId"
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "400": { $ref: "#/components/responses/BadRequest"}
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": {$ref: "#/components/responses/InternalServerError"}
      
    delete:
      tags: ["post"]
      operationId: unlikePhoto
      summary: Unlike a post.
      description: |-
        Unlike a post.
      parameters: 
        - $ref: "#/components/parameters/PostId"
      responses:
        "204": { $ref: "#/components/responses/NoContent" }
        "400": { $ref: "#/components/responses/BadRequest"}
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": {$ref: "#/components/responses/InternalServerError"}

    get:
      tags: ["post"]
      operationId: getLikes
      summary: Gets the list of user that liked the post.
      description: |-
        Gets the list of user that liked the post. The user that blocked the requester will not be shown.
      parameters: 
        - $ref: "#/components/parameters/PostId"
      responses:
        "200":
          description: Returns the list of users that liked the post.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserList" }
        "400": { $ref: "#/components/responses/BadRequest"}
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": {$ref: "#/components/responses/InternalServerError"}
...
components:
  securitySchemes:
    bearerAuthentication:
      description: |
        User token that uses the entityId of the logged user.
        Done accordingly to the project specification
      type: http
      scheme: bearer    
  schemas:
...
    UserReduced:
      description: |
        The Users entity with only the essential properties.
      type: object
      properties:
        userId:
          type: integer
          example: 1
        username:
          type: string
          example: "Username"
    UserList:
      type: array
      description: List of Users
      items: 
        $ref: "#/components/schemas/UserReduced"
...
  parameters:
...
    PostId:
      in: path
      name: postId
      description: The EntityId of the post.
      schema:
        type: integer
        example: 1
        maximum: 18446744073709551615 
        minimum: 0
      required: true
...
  responses:
...
    NoContent:
      description: The request ended successfully with no body
    Unauthorized:
      description: The access token is missing or it's expired
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs.
...